+++ src/main/java/com/hm/achievement/listener/AchieveXPListener.java
@@ -30,7 +30,7 @@ public class AchieveXPListener extends AbstractListener implements Listener {
 			return;
 		}
 
-		int previousMaxLevel = plugin.getPoolsManager().getAndIncrementStatisticAmount(category, player, 0);
+		int previousMaxLevel = (int) plugin.getPoolsManager().getAndIncrementStatisticAmount(category, player, 0);
 
 		if (event.getNewLevel() <= previousMaxLevel) {
 			return;
